requirements.txt
asgiref==3.7.2
colorama==0.4.6
coverage==7.3.0
Django==4.2.4
django-cors-headers==4.2.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
factory-boy==3.3.0
Faker==19.3.1
flake8==6.1.0
flake8-html==0.4.3
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.1
phonenumberslite==8.13.17
pluggy==1.2.0
psycopg==3.1.9
psycopg-binary==3.1.9
pycodestyle==2.11.0
pyflakes==3.1.0
Pygments==2.15.1
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
